Add files via upload
This is the Juypter Notebook file that has the instructions that were given to me and my solutions to every step of a problem. The overall gist of the project is coming up with a linear regression solution to estimate the best line of best fit given the linear model. The linear model describes the maximum height that a ball is able to reach with a bounce height (x).